fix: market resolution refactor
