+++ x/market/client/cli/tx_market.go
@@ -8,23 +8,20 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// CmdAddEvent CLI registration for add market command
-func CmdAddEvent() *cobra.Command {
+// CmdAddMarket CLI registration for add market command
+func CmdAddMarket() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "add [ticket]",
 		Short: "create new market",
 		Long:  "Create a market with ticket.",
 		Args:  cobra.ExactArgs(1),
-		//Example: fmt.Sprintf("%s tx %s postprice bnb:usd 25 9999999999 --from validator",
-		//	version.AppName, types.ModuleName),
-		// TODO: Put example
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err
 			}
 
-			msg := types.NewMsgAddEvent(
+			msg := types.NewMsgAddMarket(
 				clientCtx.GetFromAddress().String(),
 				args[0],
 			)
@@ -40,21 +37,20 @@ func CmdAddEvent() *cobra.Command {
 	return cmd
 }
 
-// CmdUpdateEvent returns query for updating market transaction
-func CmdUpdateEvent() *cobra.Command {
+// CmdUpdateMarket returns query for updating market transaction
+func CmdUpdateMarket() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "update [ticket]",
 		Short: "update market",
 		Long:  "Update a market with ticket.",
-		// TODO: put update example
-		Args: cobra.ExactArgs(1),
+		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err
 			}
 
-			msg := types.NewMsgUpdateEvent(
+			msg := types.NewMsgUpdateMarket(
 				clientCtx.GetFromAddress().String(),
 				args[0],
 			)
