+++ x/market/client/cli/tx_resolve_market.go
@@ -8,21 +8,20 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// CmdResolveEvent registers the resolve market command
-func CmdResolveEvent() *cobra.Command {
+// CmdResolveMarket registers the resolve market command
+func CmdResolveMarket() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "resolve [ticket]",
 		Short: "set resolution of an market",
 		Long:  "Resolve a market with ticket.",
-		// TODO: Put example
-		Args: cobra.ExactArgs(1),
+		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
 				return err
 			}
 
-			msg := types.NewMsgResolveEvent(
+			msg := types.NewMsgResolveMarket(
 				clientCtx.GetFromAddress().String(),
 				args[0],
 			)
