+++ x/market/keeper/export_test.go
@@ -9,16 +9,16 @@ import (
 // to export unexported methods of the keeper
 type KeeperTest = Keeper
 
-func (k KeeperTest) ValidateEventResolution(resolutionPayload types.MarketResolutionTicketPayload) error {
+func (k KeeperTest) ValidateMarketResolution(resolutionPayload types.MarketResolutionTicketPayload) error {
 	return resolutionPayload.Validate()
 }
 
-func (k KeeperTest) ValidateEventAdd(ctx sdk.Context, addPayload types.MarketAddTicketPayload) error {
+func (k KeeperTest) ValidateMarketAdd(ctx sdk.Context, addPayload types.MarketAddTicketPayload) error {
 	params := k.GetParams(ctx)
 	return addPayload.Validate(ctx, &params)
 }
 
-func (k KeeperTest) ValidateEventUpdate(ctx sdk.Context, updatePayload types.MarketUpdateTicketPayload, previousEvent types.Market) error {
+func (k KeeperTest) ValidateMarketUpdate(ctx sdk.Context, updatePayload types.MarketUpdateTicketPayload, currentMarket types.Market) error {
 	params := k.GetParams(ctx)
 	return updatePayload.Validate(ctx, &params)
 }
