+++ x/market/types/message_market_resolve.go
@@ -5,30 +5,30 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
-const typeMsgAddEvent = "add_market"
+const typeMsgResolveMarket = "resolve_market"
 
-var _ sdk.Msg = &MsgAddMarket{}
+var _ sdk.Msg = &MsgResolveMarket{}
 
-// NewMsgAddEvent creates the new input for adding an market to blockchain
-func NewMsgAddEvent(creator string, ticket string) *MsgAddMarket {
-	return &MsgAddMarket{
+// NewMsgResolveMarket accepts the params to create new resolution body
+func NewMsgResolveMarket(creator, ticket string) *MsgResolveMarket {
+	return &MsgResolveMarket{
 		Creator: creator,
 		Ticket:  ticket,
 	}
 }
 
 // Route return the message route for slashing
-func (msg *MsgAddMarket) Route() string {
+func (msg *MsgResolveMarket) Route() string {
 	return RouterKey
 }
 
-// Type returns the msg add market type
-func (msg *MsgAddMarket) Type() string {
-	return typeMsgAddEvent
+// Type return the resolve market type
+func (msg *MsgResolveMarket) Type() string {
+	return typeMsgResolveMarket
 }
 
 // GetSigners return the creators address
-func (msg *MsgAddMarket) GetSigners() []sdk.AccAddress {
+func (msg *MsgResolveMarket) GetSigners() []sdk.AccAddress {
 	creator, err := sdk.AccAddressFromBech32(msg.Creator)
 	if err != nil {
 		panic(err)
@@ -37,13 +37,13 @@ func (msg *MsgAddMarket) GetSigners() []sdk.AccAddress {
 }
 
 // GetSignBytes return the marshalled bytes of the msg
-func (msg *MsgAddMarket) GetSignBytes() []byte {
+func (msg *MsgResolveMarket) GetSignBytes() []byte {
 	bz := ModuleCdc.MustMarshalJSON(msg)
 	return sdk.MustSortJSON(bz)
 }
 
-// ValidateBasic validates the input creation market
-func (msg *MsgAddMarket) ValidateBasic() error {
+// ValidateBasic validates the input resolution market
+func (msg *MsgResolveMarket) ValidateBasic() error {
 	_, err := sdk.AccAddressFromBech32(msg.Creator)
 	if err != nil {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
@@ -52,5 +52,6 @@ func (msg *MsgAddMarket) ValidateBasic() error {
 	if msg.Ticket == "" {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "invalid ticket param")
 	}
+
 	return nil
 }
