+++ x/market/types/ticket.go
@@ -15,10 +15,10 @@ func (payload *MarketAddTicketPayload) Validate(ctx sdk.Context, p *Params) erro
 		return err
 	}
 
-	if payload.Status != MarketStatus_MARKET_STATUS_ACTIVE &&
-		payload.Status != MarketStatus_MARKET_STATUS_INACTIVE {
+	if !(payload.Status == MarketStatus_MARKET_STATUS_ACTIVE ||
+		payload.Status == MarketStatus_MARKET_STATUS_INACTIVE) {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "acceptable status is active or inactive")
-	} // TODO: Should it not be or check with reverse states
+	}
 
 	if !utils.IsValidUID(payload.UID) {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "invalid uid for the market")
@@ -74,10 +74,10 @@ func (payload *MarketAddTicketPayload) Validate(ctx sdk.Context, p *Params) erro
 // Validate validates market update ticket payload.
 func (payload *MarketUpdateTicketPayload) Validate(ctx sdk.Context, p *Params) error {
 	// updating the status to something other than active and inactive
-	if payload.Status != MarketStatus_MARKET_STATUS_ACTIVE &&
-		payload.Status != MarketStatus_MARKET_STATUS_INACTIVE {
+	if !(payload.Status == MarketStatus_MARKET_STATUS_ACTIVE ||
+		payload.Status == MarketStatus_MARKET_STATUS_INACTIVE) {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "supported update status is active or inactive")
-	} // TODO: Should it not be or check with reverse states
+	}
 
 	if err := validateMarketTS(ctx, payload.StartTS, payload.EndTS); err != nil {
 		return err
@@ -96,9 +96,9 @@ func (payload *MarketUpdateTicketPayload) Validate(ctx sdk.Context, p *Params) e
 // Validate validates market resolution ticket payload.
 func (payload *MarketResolutionTicketPayload) Validate() error {
 	// resolution status should be canceled, aborted or result declared
-	if payload.Status != MarketStatus_MARKET_STATUS_CANCELED &&
-		payload.Status != MarketStatus_MARKET_STATUS_ABORTED &&
-		payload.Status != MarketStatus_MARKET_STATUS_RESULT_DECLARED {
+	if !(payload.Status == MarketStatus_MARKET_STATUS_CANCELED ||
+		payload.Status == MarketStatus_MARKET_STATUS_ABORTED ||
+		payload.Status == MarketStatus_MARKET_STATUS_RESULT_DECLARED) {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "resolution status passed for the market is invalid")
 	}
 
@@ -128,6 +128,34 @@ func (payload *MarketResolutionTicketPayload) Validate() error {
 	for _, wid := range payload.WinnerOddsUIDs {
 		if !utils.IsValidUID(wid) {
 			return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "odds-uid passed is invalid")
+		}
+	}
+
+	return nil
+}
+
+// ValidateWinnerOdds validates market resolution ticket payload winner odds.
+func (payload *MarketResolutionTicketPayload) ValidateWinnerOdds(market *Market) error {
+	if payload.Status == MarketStatus_MARKET_STATUS_RESULT_DECLARED {
+		if payload.ResolutionTS < market.StartTS {
+			return ErrResolutionTimeLessThenStartTime
+		}
+
+		validWinnerOdds := true
+		for _, wid := range payload.WinnerOddsUIDs {
+			validWinnerOdds = false
+			for _, o := range market.Odds {
+				if o.UID == wid {
+					validWinnerOdds = true
+				}
+			}
+			if !validWinnerOdds {
+				break
+			}
+		}
+
+		if !validWinnerOdds {
+			return ErrInvalidWinnerOdds
 		}
 	}
